bi-gcp-application/Cloud_Functions/Collections/requirements.txt
# Function dependencies, for example:
# package>=version
functions-framework==3.*
pytz==2024.1
google-cloud-bigquery==3.25.0
google-cloud-storage==2.17.0
google-cloud-logging==3.10.0 
db-dtypes==1.2.0
# google-cloud-bigquery-storage
pandas==2.2.2